Add description and comments for plans for interaction between comicsplit and OCR.
Heretofore the plan was to separate each individual frame into its own file to be read by the OCR; however, such an approach is wasteful of disk space, time, and API requests. The OCR provides the x,y position of text, so this will be used to match the text with its corresponding frame.