Add find_textboxes() method to use for OCR filter
OCR sometimes identifies background text or non-text patterns within the
image.

To help avoid this, take advantage of the white-background text boxes
and use that to filter for correct text locations, as opposed (or in
addition) to filtering for within text frame boundaries.

Object method find_textboxes() uses a similar approach to find_frames,
with adjusted threshold values and area limits.